requirements.txt
google-generativeai>=0.8.3
google-cloud-alloydb-connector
gradio
sqlalchemy
pg8000
cloud-sql-python-connector[pg8000]
